Disallow live selection feathering on XP
Performance is abysmal and not easily fixed, and the new Select ->
Feathering menu provides a fine alternative.

Live feathering up to 100px is still available in Vista/7, though
performance remains poor in Vista and I have no intention of fixing it.
(Vista is effectively "not officially supported" at this point.)