Select -> Border complete
With this commit, all Select menu items are now operational, all are
properly tied into Undo/Redo, all can saved/loaded, and all are
future-proofed against ANY type of selection.  This means that when
magic wand, polygon, and freehand selections are added, they will work
out-of-the-box with all Select menu items with no extra coding.

Thank goodness for that.