Many Curves tool interface improvements
This has become quite a lovely tool, perhaps the most lovely one in all
of PD.  The histogram overlay is particularly nice, and a welcome
addition to the tool.  I'd love to add a similar feature to other tools
in PD.

With this commit tool, the Curves tool is mostly done.  I may add an
option for a diagonal guide in the morning, but other than that, this
one should be considered ready for prime-time.