Don't auto-adjust zoom after a resize operation (image or canvas)
I've gone back and forth on this behavior a number of times, and I'm now
returning to the accepted standard of NOT changing zoom after a resize
operation.  I've found it confusing when doing something like doubling
or halving the size of an image, but the auto-zoom results in an
on-screen image that looks like it hasn't changed size at all.

In cases like this, better to be predictable than clever, I think!