New pdXML class for reading/writing/parsing XML files
PhotoDemon already makes use of XML in two places: the translation
engine, and the metadata engine.  These are both highly specialized
implementations; the translation engine because it must be very fast (as
it parses thousands of phrases on any given run), and the metadata
engine because it interacts with a 3rd-party plugin with its own quirks.

In writing those two XML interfaces, it became obvious that PhotoDemon
could greatly benefit from a general-purpose XML read/write/parse
engine.  This would allow me to move all custom filetypes (filters,
macros, etc) to human-readable, non-binary files, which would make it
much easier to add new features in the future without breaking old
files, as the old engines could simply ignore any tags they don't
recognize - vs the current binary files, which must be byte-perfect or
the program will crash.

This new pdXML file is the result of my work.  It is meant to be a
lightweight XML interface that any function can use to read and write
its own custom XML data to file.  All XML files written with this class
will include a root <pdData> tag to mark the contents as
PhotoDemon-compatible.

I used the Custom Filter load/save function to test the class, which
means that it currently supports only a subset of desired features - but
the features it has are carefully debugged, optimized, and commented.

Note that this class will NEVER be a full-fledged XML parser, as I only
require a small subset of XML features for PD.  That said, the XML files
it generates will always be valid, and because I'm a bit OCD, the files
it writes will also be nicely indented according to standard XML rules.
:)