Convert Custom Filter saved files to XML
Via new pdXML class.  Old custom filters are no longer supported as of
this commit - sorry!  There are no plans to add backwards compatibility.

Also included in this commit are the two files I used to test the
feature.  They're nothing special, but I figure they may be useful to
others who want to play with the new pdXML class.