When written to file, image layer data can now be compressed
This feature is turned ON by default.  Technically it should be tied
into a user preference, as compression does slow Undo/Redo processing,
but with large HDD savings.  Hard to say that one setting is preferable
over another, but I will wait to see if I get any negative feedback over
the change.  If I don't, I may forgo a new preference in the name of
simplicity.

As a demonstration of the improvement this offers, my test raster
selection file went from 1522kb to 9kb.  Pretty good savings, eh?  :)