Update zLib to 1.2.8
zLib uses C calling conventions (CDECL) by default.  A WINAPI (STDCALL)
variant is required for use in classic VB.  In the past, Gilles Vollant
provided a precompiled download of this at
http://www.winimage.com/zLibDll/, but that version has not been updated
since 1.2.5.  The current zLib version is 1.2.8 and it includes a number
of fixes and improvements.

So I compiled it myself.  A few notes for future compilations:

- /safeseh needs to be manually added to the bld_ml32.bat file in the
/contrib/masmx86 folder.
- The VERSION 1.2.8 line needs to be rewritten as VERSION 1.28.