All paramString-based selection operations now use relative coordinates
This means that all selection operations recorded as part of a macro use
RELATIVE coordinates (instead of the previous ABSOLUTE coordinate
system).  This makes selections much more useful during macro recording.
For example, if you record a macro on a 100x100 image that includes the
action of selection the top half of the image (0,0-100,50), then play
back that macro on a 200x200 image, the selection will automatically be
updated to (0,0-200,100).

This change means that old selection files probably won't work, so I've
removed them from the repo.

I have tested this update to the best of my ability and found no issues,
but additional testing should be performed on more complex selection
actions (preferably via batch processing).