Convert the main user preferences file to XML format
Warning: huge commit!  (Most of the files in this commit have changed
simply due to method renaming and improved comments, but a few received
large code changes.)

PhotoDemon's .ini file is now officially obsoleted.  In its place, a new
preferences XML file is used.  The new XML file is written, read,
updated and parsed via the pdXML class, which has received a number of
functions to help it fulfill all the needs required by an INI-like
handler.

I have gone back and forth on whether to convert old .ini files to the
new .xml format.  It's a fair amount of code to do this, and where 6.0
will break backward compatibility in so many ways, I feel I may as well
just go full-bore and ignore the old .ini file completely.  I don't
think it worth the trouble to write special code just for this,
especially where the preferences can be easily reset.

If anything, performance is better under the new system.  Windows INI
API calls have been obsolete for years.  Actually, I quite like the
warning on the WriteProfileString dev page: "Note  This function is
provided only for compatibility with 16-bit versions of Windows.
Applications should store initialization information in the registry."

Some INI code remains because the PhotoDemon updater relies on an INI
file for update information.  I will be changing this to XML as well for
the 6.0 release, at which point INI interactions will be completely
removed from the software.