Rewrite exposure tool algorithm
Relevant to #74.

After much research, I have settled on a new exposure formula... a
surprisingly simple formula, actually.  Exposure compensation is now
calculated on basic power-of-2 scale, per the standard "stop" definition
(http://en.wikipedia.org/wiki/Exposure_value).

Effectively, this means that each exposure compensation of -1 cuts the
image's luminance in half, while each exposure compensation of +1
doubles total luminance.  This is the behavior of other tools
(http://www.digikam.org/ in particular was used as a reference), and it
also mimics the behavior of an on-camera manual EV adjustment.

I still find it useful to display the new EV curve, even though it's not
a curve, as it accurately demonstrates the luminance change.  EV
adjustments will result in shadow or highlight clipping; that is
intentional, and could even be useful for an eventual HDR tool (although
I may still just use Unsharp Mask for that... will revisit later).