Add uncompiled resources to repository
This is a bit of a tricky decision, as many of the images in
PhotoDemon's resource file are not created by me, and are thus governed
by their own licenses.  I have added a very explicit LICENSE and README
file to the folder to help others understand this.

That said, it will be very nice to have these items in the repository.
I frequently tweak the various menu and button png files, and this way I
can finally track those changes.  It will also be nice to have a tracked
copy of the resource script I use for the primary resource file.