Fix "Fade Last Effect" not enabling correctly
This item really needs a more elegant system for determining when to
en/disable.  It can be triggered from multiple places - making it tough
to debug - and the new "selections being part of Undo/Redo" makes it
disable when a selection is created, when really it should still operate
on the last valid pixel-based action.

I've made a note and will investigate this further at some point.