Fix ClearType initializing based on RR research
Robert has helped me track down a failure in PD's initialization of
ClearType antialiasing for PD text.  The API call requires an explicit
ByVal for a PVOID param - now, PD can enable ClearType for the duration
of the program if it is not already enabled.  This improves readability
greatly on XP.