Remove VB binary file import feature
While I find this feature clever, I no longer have a direct use for it,
and two classes, a module, and a form is overkill for such a fringe
item.

Also, I'm trying to cut as much cruft as possible before the 6.0
release.  This is one of those items that makes way more sense as a
plugin, not as a core program feature, so perhaps I will revisit it in
the future.

For those who still need this ability, old versions of PD will
(obviously) support it.