New tool dialog "Command Bar" custom control
Historically, I have simply copy+pasted OK and Cancel buttons to each
new tool dialog.  I will shortly begin removing all those buttons in
favor of this new, universal "command bar" control, which includes not
only OK and Cancel buttons, but also a "reset" button (to reset all
controls to default values) and soon, save/load presets, and possibly a
"randomize" button a la Paint Shop Pro.

Because this command bar will be used on every tool dialog, I am
wrapping a number of functions into it that had to be previously be
handled uniquely by each form.  For example, the command bar will
automatically unload the parent form on cancellation - no need to
rewrite that code for every dialog.  When OK is pressed, all text-based
controls will automatically be validated, and if all validation checks
pass, the parent form will automatically be hidden from view prior to
executing the relevant processing statement.  If validation checks fail,
a message will be displayed, and the relevant control will be
highlighted.  Stuff like that.

This is going to cut quite a bit of redundant code from the program, as
well as paving the way for #67, #68, and #75 to finally be implemented.
I'm quite excited about it.