Another big Command Bar update
As of this commit, this control *almost* resolves any remaining issues
with #75, and it also effectively removes all blockers from #67 and #68.

Highlights of this commit include:
- Reset to Default behavior is now automated, with a hook for dialogs
that need to do their own work.  After additional pondering, I decided
it was worthwhile to automate this process, as most dialogs will not
need highly specialized "default" values.
- Randomize is now implemented.  It wasn't much work, and it's fun for
playing with unfamiliar tools, so I went ahead and did it
- User preset loading/saving works FOR ALL STANDARD CONTROLS.  I still
need to write custom hooks for non-standard dialogs (like Curves)
- Last-used settings are successfully saved/written to file

Also of mention - I spent a frightening chunk of time trying to
implement "press Enter to save" behavior for the preset combo box, but
was unsuccessful.  Because the command bar has a Default button ("OK"),
VB eats all Enter keypresses and there is no way to reclaim them.
Dynamically unsetting OK as the Default button does not resolve the
issue, presumably because VB does not re-check the presence of a Default
button after the form is loaded.  I will eventually subclass the combo
box and check the keypresses myself, but in the meantime, you must
actually click the Save button to save the current preset.