New color selector user control
Unfortunately, this is not a full-blown "color selector".  This is
simply a thin user control to replace the picture boxes I was previously
using for color selection on tool dialogs.

The advantage of a user control (besides less copy/paste code throughout
the project) is that it interacts nicely with the new Command Bar.  The
color selector can be easily reset, randomized, and saved/loaded as a
preset.