Fix command bar to work with control arrays
If a control is part of a control array, we can't simply store its name,
as it shares that same name with other control!  We must also store its
index, then retrieve that index at read time.

Also fixes a bug with resetting according the Min property of certain
controls