Add command bar to more tools
Gonna be a lot of commits like this.

Channel Mixer proved to be a nasty project.  It retains a ton of
off-screen information, which the command bar (obviously) can't handle.
So getting it to load/store all hidden values and enable/disable
controls as necessary proved difficult.  I believe it is now working
without incident, but I've made a note to do some additional testing
before launch.