Move the old toolbar module into Interface, and rename old functions
In way old versions of PD, the toolbar module and accompanying tInit
function were used to enable/disable multiple controls based on meta-
labels, e.g. "Undo" which requires a number of menu items and toolbox
buttons to be triggered.

Because I no longer have a toolbar, and tInit is not very descriptive,
I've moved the code into the more-appropriate Interface module and
renamed associated functions to match.