Separate various "add transparency" options
Placing so many transparency options on the same dialog was a dumb idea.
PD tries to adhere to the "do one thing and do it well" rule for tool
dialogs, so I've separated out the "green screen" option into its own
tool.  This also makes it far more discoverable.

A few other items in this commit.  I now allow "add alpha channel" tools
to be used on 32bpp images, as it may be useful to modify an existing
alpha channel.  "Remove alpha channel" is still disabled for 24bpp
images.

Also I've rewritten the Image -> Transparency menu to exclusively use
"transparency" instead of "alpha".  This is in keeping with the Options
dialog as well.

Finally, both add alpha dialogs have been converted to use the command
bar.