Rewrite central convolution implementation (DoFilter) against param strings
I put this off for a long time because it was an unpleasant task, but
I'm happy to say it is now complete.  This rewrite was required before
updating the custom convo filter dialog to use the new command bar.

A number of dialogs and functions had to be rewritten to support this -
edge detection, sharpen, and a few obscure ones.  Hence the large size
of the commit.