New Rotate distort
I needed an internal rotate function separate from the full-image rotate
tool, because it has to operate on selections.  It's not practical to
merge the two tools into one because the full-size one must be capable
of enlarging the canvas, while the selection one cannot.  Also I needed
a rotate function that supports PD's full distort engine, with various
edge-wrapping filters, for a radial blur function.