New Radial Blur (spin and zoom) tool
This is my first stab at a radial blur tool.  Radial blur (both types)
presents a serious computational problem.  Blur operations are always
slow (because each pixel requires you to process many others), and
radial blur compounds the problem by not allowing you to re-use the
results of previous pixels.

I've opted for an approach where you convert the image to polar
coordinates, apply the blur, then convert back.  This has trade-offs.
It's pretty fast, which is good, and it can use an optimized box blur
for the blur itself - also good, because PD has one of those.  But it's
not great from a quality standpoint, because the rect -> polar -> rect
conversion causes a measurable loss of quality.

I'm not sure how other software overcomes this.  I'm going to do some
additional research, but in the meantime, the tool is at least ready for
casual use.