Vastly improved radial blur
Combining zoom and radial blur into a single function saved me some
programming work, but I don't think it's a sustainable long-term
investment.  Alas.  I have therefore excited the "zoom" option from the
radial blur tool, and will investigate a custom implementation that
doesn't suffer the same issues as this one.

This commit also includes a fix for the center-line blur issue
previously noted, which means standalone radial blur is now quite
lovely.  Performance is an issue; I'm going to resolve this with a
custom horizontal blur implementation.  The current box blur is designed
around the assumption of both horizontal and vertical blur actions,
which is great for standard box blur, but not great for horizontal blur
alone.  I could also use an optimized horizontal blur for the
forthcoming motion blur tool, so it's a worthy investment.

(Aside from motion and zoom blur, I think I am mostly good-to-go on the
5.6 new tool list.  Yay!)