Convert custom rotate code to a standalone function...
...that operates on a source and destination layer.  This is required
for the forthcoming motion blur, which requires access to the newest
custom rotate function.

Also, this commit includes a rather odd function that grows a layer and
clamps all new borders to the value of the nearest valid pixel.  This is
required for certain functions (again, forthcoming motion blur) that
require access to off-image pixels.  Rather than write complicated
edge-handling code, it's easier to just up-size the image, perform the
function, then downsize to original dimensions.