New Zoom Blur tool
This is a rough implementation using nothing more than StretchBlt and
AlphaBlend.  The end result is very fast, but with less nuanced results
than comparable tools in other photo editors.  I'm going to do a bit of
extra research to see if I can mimic their results while maintaining the
speed advantages of this implementation, but in the meantime, the tool
is ready for testing.