Provide a "reverse" option in the Perspective Distort tool
This brings it into parity with GIMP's perspective tool.

I previously avoided adding this option because it's a confusing
distinction for the user (seriously, GIMP presents it terribly).  But I
think I've found a way to present it that will make sense to casual
photographers who just want to fix a photograph's orientation, without
overwhelming them.

Also, the "reverse" option is quite helpful if the image contains a
reference object that needs to be converted from a quadrilateral to a
rectangle.