Fully implement metadata load preference (and delayed metadata parsing)
The user can now choose to disable load-time metadata parsing.  If
selected, metadata will not be parsed until it is requested by one of
two operations: saving the image to file (with a metadata export option
enabled), or initiating the metadata browser.

With this option enabled, there are a few issues.  Because we no longer
know in advance if an image has metadata, the metadata browser is
enabled by default.  If the user clicks it, metadata will be loaded then
- and if none is found, the user will be warned and the metadata browser
will cease to load.

We also don't know in advance if GPS metadata is present.  This option
is turned off by default, but will be re-enabled if the metadata browser
is initiated and GPS data is found.

This commit also improves metadata export behavior if the preference is
set for not exporting metadata at all.  If this preference is chosen,
the entire metadata check process at export time is now skipped - so
users can effectively disable all metadata loading and saving if
desired.  (Even with both load and save metadata turned off, metadata
can still be browsed on-demand by clicking the Image -> Metadata menu.)