Display wait screen while changing program language
This was needed to prevent interaction while the new language is being
loaded.  A marquee progress bar is used, but it will look funny in the
IDE due to the lack of a manifest.  (It looks great when compiled,
especially on Vista+.)

The new "please wait" form is generic and can be used by any function
where an indeterminate progress bar is required.