More Language Editor features
All phrases are now internally saved (but not written to file!  Don't
use this tool for actual work yet!).  Shortcut key implemented for very
fast translating.  Number of phrases in the current list now displayed.
Unsaved translations displayed in red.

Next I will need to add a function to the XML engine to replace a tag at
a generic location, and ideally with knowing only the tag name and not
its contents.  (I've rigged the phrase type to include the original
translation in case this is impractical, but I think it'd be better to
drop this and simply remove the tag and its contents intelligently.
Shouldn't be too tricky to code, but will ponder a bit before committing
to either approach.)