New save/load last-used settings class
This is basically a trimmed-down version of the command bar's save/load
preset code, with a few modifications to make it work as a class.

The neat thing about having this available as a class is that any form
can utilize one, and all controls on that form will automatically have
their values saved/loaded to file upon demand (with hooks for custom
data as well!).  In the future, this would allow me to perfectly restore
a full batch wizard state, for example.