If GDI+ is available, use it to accelerate 32bpp viewport rendering
The long-term goal is to move multiple elements of the viewport rendered
over to GDI+ where appropriate.  If I'm ever to add layers (or even
paint tools) to PD, I'll need something faster than pure VB can provide,
and basic GDI is far too unreliable with 32bpp data to use for most
functions.

As a test bed, I've handed over 32bpp image stretching to GDI+ when the
viewport is zoomed out.  Performance with basic bilinear stretching is
excellent, and GDI+ handles the alpha channel properly, which saves us
having to do it ourselves!