During an effect preview, shift the preview ratio burden to prepImageData
Previously, functions that operate on a radius (like blur) had to
perform their own math when rendering a preview, to make it so that the
previewed blur strength was correctly representative of the final blur
strength.

This was a foolish design decision.  prepImageData is perfectly capable
of calculating this itself, and it now does so.  Now, calling functions
can simply borrow the calculated data as necessary.

Also, greatly improved comments in this module with this commit.