Make histogram form modal, plus other optimizations/changes
I like the thought of having the histogram be non-modal, so the user can
visualize histogram changes as they happen, but this leads to a bunch of
unforeseen interactions that the program does not currently handle well.
(The histogram is regenerated with alarming frequency - this could be
avoided by storing histogram data inside each pdImage object, but that's
a large change, and I don't want to do that until I have a dedicated
histogram class available.)

Rather than add a bunch of new histogram stuff this close to release,
I've added the issue to the 6.2 roadmap.

Also, the histogram form should now be much better optimized.
Previously, double-redraws were being accidentally fired, and a lot of
old, sloppy code was being used.  All fixed by this commit.