Add new "fast" option to Gaussian blur; 20x speed improvement!
So, the math is a bit involved (this link explains it well:
http://nghiaho.com/?p=1159), but basically, a Gaussian function can be
estimated by iterating multiple box functions.  In the case of Gaussian
blur, a Gaussian blur of radius 3 is matched within 3% (Central Limit
Theorem!) by three Box blurs of size 1.

Because Box blurs are significantly easier to optimize (see my two
previous commits), such an estimation is generally an acceptable
trade-off, especially for CPU-bound blur operations like PD's.

As such, PhotoDemon now provides this estimation as an option.  The
speed-up is huge: ~20x, on average, in my testing, with bigger gains as
the radius increases.  This matches the processing speed of GIMP, and
far surpasses Paint.NET.  Very cool.

Note that there are trade-offs to using the fast mode.  Only
integer-radii are supported, and the blur is slightly lower quality.
(Lower quality could be theoretically solved with more iterations, but
at a speed trade-off.  PD's core Gaussian blur also uses a modified
sigma calculation for improved results, so even additional iterations
will not exactly reproduce the true Gaussian.)

For most purposes these losses shouldn't matter, especially in the face
of a 20x speed-up.  (But the original Gaussian is still available for
those who want/need it.)