Fill in some additional hotkeys
(FYI - as this is low-impact stuff, I don't mind doing it this close to
release!)

Hotkeys are an interesting usability issue.  I tried to find data on how
many users use hotkeys (but couldn't), but I imagine the number is
extremely low, outside of perhaps Ctrl+O, Ctrl+V, etc.

Hotkey usability is closely tied to how reflexive they are, so I have
added a few more hotkeys that match Photoshop's.  Casual users will not
even notice these, but power users should appreciate it.

The plan for 6.2 is to give the user a way to assign hotkeys to
arbitrary actions.  This is the only way to *really* make hotkeys
usable, as different users have different needs, and it is not feasible
to assign hotkeys to every menu in the program.

To improve usability, I'm also going to add an option to "match
Photoshop" or "match GIMP" hotkey layout.  For power users, this will
save them having to prepare their own elaborate keymappings.

(For this release, however, the current mash-up of Photoshop and
PD-specific hotkeys will have to do.)