Reorganize Adjustments menu
(Add this to the very long list of "things I wish I'd thought of BEFORE
releasing a beta")

WHAT PROMPTED THIS: two things, really:

1) As I look through my 6.2 roadmap, I see a number of new additions to
the Adjustments menu.  Many of these are "auto" versions of existing
tools - auto white balance, auto contrast, etc.  There are also a few
outliers I don't currently offer, like split-toning, gradient map,
variations, or a dedicated HDR tool.

2) One of the things I really want to hit in the 6.2 release cycle is
help documentation.

THE PROBLEM: PD's current Adjustments menu is *huge*.  It's a rough
mash-up of Photoshop and GIMP, which is to say that it's not organized
well.  A lot of tools were added in the 6.0 release cycle, and while the
current organization makes sense to me, I doubt it's immediately
apparent to anyone else!

The addition of a bunch of new tools would only compound the problem,
and I'll have no choice but to eventually rework it into something more
manageable.

In addition: if I'm going to be writing help documentation with loads of
screenshots, I want the interface as set in stone as possible.  It is
incredibly time-consuming to redo a bunch of documentation because my
menu layout changed, so 6.0 needs to have menus as finalized as
possible.

THE SOLUTION: retool the Adjustments menu now, rather than later.  It's
horrible for me to be introducing changes like this in the midst of a
beta, but better now than after all the 5.4 users get an update
notification!

The new organization is quite nice, I think, and discoverability should
actually be improved by the redesign.