Next-to-last language file update before release
..."in theory".  I must always preface my commits with "in theory".  ;)

The final language commit will happen when I receive the translators'
final XML copies.  There are ~30 translations left to do, a result of
finding text that wasn't caught by the language file generator, new bug
fixes, and a few small items to unify text between the program and
photodemon.org.

As usual, this batch of language file updates led to minor updates
elsewhere in the program, mostly attempts to standardize language and
not use more phrases than I have to.

The master translation generator also saw a number of fixes to work
around some intricacies of the latest batch of translation engine fixes.