Remove all MDI code/objects/etc from the project
(Disclaimer: this is an inevitably huge commit that includes a number of
other items.  Sorry!)

Because there's no *good* way to remove MDI from an MDI-based project,
this commit is a sort of "halfway point".  PD still runs, and most
everything should still work (minus some oddities in the Windows menu,
and some trouble auto-zooming and positioning image forms), but we're
mostly able to mimic the original MDI behavior with non-MDI forms.

(This approach isn't all bad, minus the obvious number of oddities that
need to be resolved.)

The most interesting aspect of this implementation, IMO, is that with
some fine-tuning, there could be merit to maintaining a multi-window
layout for the program.  It's kind of like having MDI without any of the
MDI limitations; in particular, the multi-monitor possibilities of this
codebase are promising.  I'll have to investigate this possibility
further.

Next up is to start separating out the bits of code that assume unique
forms for each child window.  Once that's been cleaned, we can abstract
out the viewport bits, so that single-window or multi-window is
effectively the same for 99% of the program - only the viewport renderer
will care about the difference, and it can adjust its behavior
accordingly.