Rename some essential public variables
Some of the oldest parts of PD include the core image tracking code
(e.g. anything relating to the pdImages array).  The code surrounding
that array is generally of poor quality, including a lot of variables
with indistinct names, which is unfortunate considering that pretty much
every aspect of the program interacts with it at some point.

This commit is an attempt to fix some of the worst offenders.  (I am
also trying to be better about using a "g_" prefix on global (public)
variables.)