New requestThumbnail function
One of the interesting challenges of implementing layers is that
external functions should no longer attempt to directly access an
image's DIB, because if multiple layers are in use, any requests for a
copy of the image (such as for a thumbnail to place in the MRU list or
taskbar) will require the compositor's help.

To that end, the pdImage class now exposes a requestThumbnail function,
which all thumbnail-related functions should use in place of directly
accessing pdImage.mainLayer.