User-facing options for show/hide individual toolboxes
(Relates to #93)

This feature had more interactions than I anticipated, but I *think*
I've worked out all the kinks.

The biggest problem is that toolbar forms must never be unloaded.
Whether docked or floating, we need to keep toolbar forms in memory so
we can properly track and sync their position - otherwise, if/when the
user re-activates a toolbar, it won't appear in the proper place.
(Also, things like Undo/Redo button en/disabling needs to be applied,
even if a toolbar is not visible, so that if it is *reactivated*, all
buttons are in the proper state.)

To that end, toolbar windows are never unloaded, only hidden.  This
should be properly handled, even for things like clicking the control
box X on a toolbar form, and the Window menu items should be
automatically synced with any visibility changes.

One thing I very much like about this change is that it allows for a
truly minimal PD interface!