Hide additional Window menu items (Cascade, etc) when images are docked
I do not like to dynamically hide menu items, but I feel this is a
worthy exception.

Because Docked mode will be the default for image windows, the user does
not need to be bombarded with a bunch of disabled menu items like
Cascade, Tile, etc.  Those options are only for advanced users - the
kind of users who will switch on floating windows, and thus be
comfortable with the Window menu in general.

(I frankly doubt more than 10% of users ever click an item in a
program's Window menu anyway, though I can't find hard data to back this
up.)